requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.26.68
botocore==1.29.68
cachetools==5.2.1
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
confuse==2.0.0
Django==4.1.6
django-celery-beat==2.4.0
django-redis==5.2.0
django-timezone-field==5.0
djangorestframework==3.14.0
google-api-core==2.11.0
google-api-python-client==2.77.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
jmespath==1.0.1
kombu==5.2.4
numpy==1.24.2
oauth2client==4.1.3
oauthlib==3.2.2
pandas==1.5.3
Pillow==9.4.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-amazon-sp-api==0.18.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
slack-sdk==3.19.5
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
xmltodict==0.13.0
